refactor(engine): fold impl shims, trim per-frame overhead
- parse frame-length header as raw bytes: msgpack decode allocated a
  Decoder per incoming frame; peekBytes returns views over copies
- skip incoming logging wrapper unless printRPC
- fold index-impl/session-impl into index/session (platform-fork relic)
- drop resetClient mobile branch: unreachable, Engine.reset() no-ops
  on mobile
- ProxyNativeTransport.reset() fails outstanding invocations so
  pre-account-switch RPC callbacks can't fire against new session
- delete unused engine/types scalar aliases (codegen stopped emitting
  them), errors.msg map, local-console messageObject variant
